propagating the last value of the running sum.
we should really propagate the previous trailing week, because we now remove
the value all at once on the 7th day of the new year.

this is not such a big deal for the 7 day running sum, but I wish to have a
30 days running sum, too, and without the above fix, it will overextimate
the available water for the whole of the month January.